Fix secondary server onboarding flow
